Implement user context validation and api/lists/ methods

+++ Web/SyncLists/sync/api/lists/views.py
@@ -4,14 +4,14 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 
-from sync.api.utils import request_body_to_dict
-from sync.models import List
+from sync.api.utils import request_body_to_dict, validate_user_list_context, get_user_context
+from sync.models import List, Task
 
 
 @csrf_exempt
 def list_id(request, l_id):
     if request.method == 'GET':
-        response = get_list_by_id(request, l_id)
+        response = get_list(request, l_id)
     elif request.method == 'PUT':
         response = edit_list(request, l_id)
     elif request.method == 'DELETE':
@@ -36,9 +36,10 @@ def create_list(request):
     if request.method == 'POST':
         try:
             request_body = request_body_to_dict(request)
+            request_body['list_owner'] = get_user_context(request)
             response = List.create(**request_body).single_to_json()
             status_code = 200
-        except (TypeError, SyntaxError):
+        except (TypeError, SyntaxError, KeyError):
             response = 'Invalid request_body:\n{0}'.format(request_body)
         except:
             response = 'Unexpected exception!\nrequest_body:\n{0}\nexception:{1}'.format(request_body,
@@ -48,11 +49,15 @@ def create_list(request):
     return HttpResponse(response, status=status_code)
 
 
-def get_list_by_id(request, l_id):
+def get_list(request, l_id):
     status_code = 400
     try:
-        response = List.get_by_id(l_id).single_to_json()
-        status_code = 200
+        list = List.get_by_id(l_id)
+        if validate_user_list_context(request, list):
+            response = list.single_to_json()
+            status_code = 200
+        else:
+            response = 'Invalid user context'
     except ObjectDoesNotExist:
         response = 'List id {0} does not exist.'.format(l_id)
     except:
@@ -64,10 +69,13 @@ def edit_list(request, l_id):
     status_code = 400
     request_body = {}
     try:
-        request_body = request_body_to_dict(request)
         list = List.get_by_id(l_id)
-        response = list.edit(**request_body).single_to_json()
-        status_code = 200
+        if validate_user_list_context(request, list):
+            request_body = request_body_to_dict(request)
+            response = list.edit(**request_body).single_to_json()
+            status_code = 200
+        else:
+            response = 'Invalid user context'
     except (TypeError, SyntaxError):
         response = 'Invalid request_body:\n{0}'.format(request_body)
     except:
@@ -80,9 +88,12 @@ def delete_list(request, l_id):
     status_code = 400
     try:
         list = List.get_by_id(l_id)
-        list.delete()
-        response = 'User id {0} has been deleted.'.format(l_id)
-        status_code = 200
+        if validate_user_list_context(request, list):
+            list.delete()
+            response = 'User id {0} has been deleted.'.format(l_id)
+            status_code = 200
+        else:
+            response = 'Invalid user context'
     except ObjectDoesNotExist:
         response = 'List id {0} does not exist.'.format(l_id)
     except:
@@ -93,8 +104,12 @@ def delete_list(request, l_id):
 def get_list_users(request, l_id):
     status_code = 400
     try:
-        response = List.to_json(List.get_by_id(l_id).get_all_users())
-        status_code = 200
+        list = List.get_by_id(l_id)
+        if validate_user_list_context(request, list):
+            response = List.to_json(list.get_all_users())
+            status_code = 200
+        else:
+            response = 'Invalid user context'
     except ObjectDoesNotExist:
         response = 'List id {0} does not exist.'.format(l_id)
     except:
@@ -103,4 +118,16 @@ def get_list_users(request, l_id):
 
 
 def get_list_tasks(request, l_id):
-    return HttpResponse('get tasks for list ' + l_id)
+    status_code = 400
+    try:
+        list = List.get_by_id(l_id)
+        if validate_user_list_context(request, list):
+            response = Task.to_json(list.get_tasks())
+            status_code = 200
+        else:
+            response = 'Invalid user context'
+    except ObjectDoesNotExist:
+        response = 'List id {0} does not exist.'.format(l_id)
+    except:
+        response = 'Unexpected exception!\n{0}'.format(sys.exc_info())
+    return HttpResponse(response, status=status_code)
